WIP: Gravity Model rewrite
