+++ normits_demand/cost/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+Created on: 05/11/2021
+Updated on:
+
+Original author: Ben Taylor
+Last update made by:
+Other updates made by:
+
+File purpose:
+
+"""
+from normits_demand.cost.cost_functions import BuiltInCostFunction
+from normits_demand.cost.cost_functions import CostFunction
