+++ normits_demand/distribution/__init__.py
@@ -1,3 +1,3 @@
 # TMS Distribution Models
 from normits_demand.distribution.external_model import ExternalModel
-from normits_demand.distribution.gravity_model import GravityModel
+from normits_demand.distribution.tms_gravity_model import GravityModel as TMSGravityModel
