+++ requirements.txt
@@ -7,3 +7,5 @@ openpyxl>=3.0.5
 pyodbc>=4.0.30
 geopandas>=0.8.1
 packaging>=20.8
+
+scipy>=1.6.2
